Move enumeration counting to trait
